UX-408 Move code into src, replace input component
