+++ packages/DatePicker/package.json
@@ -13,9 +13,10 @@
   },
   "dependencies": {
     "@babel/runtime-corejs2": "^7.3.1",
-    "@paprika/icon": "^0.0.1",
-    "@paprika/popover": "^0.0.1",
-    "@paprika/tokens": "^0.0.1",
+    "@paprika/icon": "^0.1.0",
+    "@paprika/input": "^0.0.1",
+    "@paprika/popover": "^0.1.0",
+    "@paprika/tokens": "^0.1.0",
     "aphrodite": "^2.3.1",
     "moment": "^2.24.0",
     "prop-types": "^15.7.2",
