+++ packages/DatePicker/stories/DatePicker.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { storiesOf } from "@storybook/react";
 import moment from "moment";
-import DatePicker from "../DatePicker";
+import DatePicker from "../src/DatePicker";
 
 storiesOf("DatePicker", module).add("DatePicker", () => <DatePicker date={moment()} format="MM/DD/YYYY" />);
